+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 * Fix usage of `:in` when defined with a `:default` option that is not a hash, to apply the default when param is not given.
 * Fix usage of `:in` incorrectly calling scopes when receiving a blank param value without `allow_blank` set.
+* Deprecate passing a String to `if` and `unless` options, in order to stop using `eval` in code.
 * Require `active_support` and `action_controller` explicitly to prevent possible `uninitialized constant` errors.
-* Add support for Ruby 2.7. (no changes required)
-* Remove support for Rails `4.1` and `4.2` and Ruby `2.1`, `2.2`, and `2.3`.
+* Add support for Ruby 2.7 and 3.0, drop support for Ruby < 2.5.
+* Add support for Rails 6.1, drop support for Rails < 5.2.
+* Move CI to GitHub Actions.
 
 ## 0.7.2